+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     CHANGE commit - Fixes [Issue #282](https://github.com/dsccommunity/StorageDsc/issues/282)
 - Updated tag from 'Dev Drive' to 'DevDrive' in manifest file - Fixes [Issue #280](https://github.com/dsccommunity/StorageDsc/issues/280)
 
+### Changed
+
+- OpticalDiskDriveLetter:
+  - Fix detection of virtual optical drives on Windows Server 2022 21H2
+    running in Azure - Fixes [Issue #285](https://github.com/dsccommunity/StorageDsc/issues/285)
+
 ## [5.1.0] - 2023-02-22
 
 ### Changed